EquallySpaceAxis: Override Igor sort order on default sortOrder parameter
The list of axis returned by Axislist is not sorted according to enabled
region via `axisEnab`. This makes it impossible to always append a new
axis to the top or bottom of a graph.

Utilizing the new SortAxisList we can now sort the axis in order of
enabled region.

Original-CommitID: 4bcd87aa248383e2befcabf83ce403de5720da67