ZeroMQ: Throw RequestInterfaceException(REQ_OUT_OF_MEMORY) in critical code paths
When a memory allocation fails in the XOP we usually get a std::bad_alloc
exception which translates to an early stop  of the currently executing
XOP function.

By catching bad_alloc early on and translating that into a
RequestInterfaceException we can reply with an appropriate message.

Original-CommitID: 1ba38fc7ea524c3b6830c79c65796c4d9c59063d