DAP_CheckSettings: Export into NWB before DAQ if requested
This has the advantage that the user does not need to wait after the very
first sweep. Additionally it also ensures that if we have to create a
sibling experiment due to low free memory, we already have an open NWB file
and thus can be sure that we use the parent PXP for deriving the NWB
filename.

Original-CommitID: f14f2c92b7263c6904b670b92de3edcc147c7288