TangoInteract: Fix invalid stimset selection
The function TI_runStimWave accepts a stimset name and derives a stimset
popup menu index from it.

This requires that the sorting of the stimset lists are equal. A non
matching sorting triggers for some cases the bug that the wrong stimset is
run with e.g. TI_runStimWave.

The bug is present since the intial introduction of these functions.
Due to the nature of the default sorting of the stimset waves this bug
could be hidden in some cases.

Original-CommitID: c91edcbcac3633275eddb5feecc0b9477e3a5ff4