FindRange: Fix an edge case with sweep rollback and the entry source types
The following recipe

- Acquire some sweeps, with TP after DAQ
- rollback the sweep counter by one
- TP
- forward roll the sweep counter by one
- acquire some sweeps

generates a labnotebook where we are not able to query the settings of the
first sweep as the TP settings from the same sweep hides it.

So this is a edge case where we have to respect the entrySourceType
already when we are looking for the last sweep range.

We do that now in FindRange, this function was also moved to
MIES_MiesUtilities.ipf as it is now specific to the labnotebook.

This bug was present since forever. But only since the introduction of the
entry source types in 1b21f99c (Bugfix: Disambiguate labnotebook entry
search, 2016-10-08) the user could trigger it.

Original-CommitID: 7ff9978f431a24362e8efadc3e86eaa426591c97