CreateTiledChannelGraph: Fix oodDAQ display with onset delay
In 1096f2b9 (CreateTiledChannelGraph: Adjust waves for onset delay for
oodDAQ view, 2017-02-28) we changed the way we handle the onset delays for
the display in dDAQ mode.

Turns out this broke onset delays larger than zero.

Revert the part which changes the start and end x coordinates as this is
wrong. The Display operation does not honour the wave offset, so we need
the shifted xRange start and end.

Original-CommitID: 49a584980b1e241b2f0099af3636b0a93ad64506