PulseAveraging: Fix display of multiple sweeps in PPA graph
In 32ec4325 (Add pulse averaging, 2017-02-09) we added support for pulse
average graphs. The extracted single pulse waves got the same wave scaling
as the channel data waves.

This was correct at that time but due to the changes in 1096f2b9
(CreateTiledChannelGraph: Adjust waves for onset delay for oodDAQ view,
2017-02-28) the dim offset of the channel data wave can now be larger than
zero. And we don't want to keep this channel offset in the single pulse
waves.

Use an explicit SetScale call with zero offset to avoid that problem.

Original-CommitID: 25425edcd8df0beba80457d87308f227afdda2e7