PA_CreateAndFillPulseWaveIfReq: Fix work-prevention logic
We have to keep track of the timestamp of the channel data wave. As e.g.
if it is changed back to the backup wave we want to trigger a new
calculation.

We now store the timestamp of the source wave to ensure that this works.

Original-CommitID: ae43533d4ee48572aa4059eb6054519e81d01d29