IH_Cleanup: Skip calling cleanup routines if possible
For some installation methods, notably the hardware independent
installations, MIES might not be the always used product.

In that case our cleanup creates the base folders, then realizes that
there is nothing to cleanup and later on leaves them as is.

It is much more clever to avoid trying to cleanup at all if the main MIES
folder does not exist. We have to use the "AsString" version here in order
to not create the datafolder.

Original-CommitID: e2de75a11d0b8e93de53cdd4f2603190058b2db6