Disband MIES_HardwareSetUp.ipf and integrate it in MIES_PanelITC.ipf
This is a long overdue cleanup. The justification that it was not a proper
module belonging into its own file is exemplified by the fact that now all
non-gui procedures could be made static, because all callers are in
MIES_PanelITC.ipf.

Original-CommitID: c160362a3c783891ca88ee907151764c3f54901b