TP MD: Push stopCollection point further to the end
Since the dawn of this project we've used the stop collection point as the
last point of the ITCDataWave which we are interested in. This is required
as the internal FIFO is always advancing during the measurement, and the
time between the stop collection point and the end of the ITCDataWave can
be used to stop the hardware.

Since the introduction of the separate fifo monitoring in 7b190f70 (TP MD:
Rework and fix crashes with 64bit XOP, 2016-06-27) for TP MD this safety
margin between stop collection point and the end of the ITCDataWave can be
reduced much further.

In DC_CalculateITCDataWaveLength we use the next power of 2 of the stop
collection point and together with a lower boundary this, for default
settings, resulted in a ITCDataWave length of
2^MINIMUM_ITCDATAWAVE_EXPONENT.

But in DC_UpdateTestPulseWaveMD we a calculation which was always slightly
above 2^MINIMUM_ITCDATAWAVE_EXPONENT therefore we ended up with a
stop collection point which was in the middle of the ITCDataWave and the
ITCDataWave had a length of 2^(MINIMUM_ITCDATAWAVE_EXPONENT + 1).

Now we move the stopCollection point to 90% of
2^(MINIMUM_ITCDATAWAVE_EXPONENT + 1) for TP MD.

Original-CommitID: b66a52245f667a8325beb7b690b809243bf061d2