Labnotebook: Add stimset checksum to textual labnotebook
Some time ago while debugging a problem it turned out that a user had
changed the contents of a stimset but not the name. It turned out to be
pretty difficult to prove that though.

Now we store the CRCs of the stimset parameter waves, or the CRC of the
stimset for third party stimsets, in the textual labnotebook as well.

Original-CommitID: 06ecbbbe1e7fcb292ee11f100774b09a34250328