PA_GetPulseToPulseLength: Fix fallback logic
In 306f99c5 (WaveBuilder/Pulse Trains: Add the pulse to pulse length to
the wave note, 2017-02-09) we added support for writing the pulse to pulse
length in the labnotebook.

But the pulse averaging code, available since 306f99c5 (WaveBuilder/Pulse
Trains: Add the pulse to pulse length to the wave note, 2017-02-09) still
expects a non existing labnotebook entry for the pulse to pulse length as
an indicator for a poisson distribution.

Change that to treat a non existing entry and a zero pulse to pulse length
(which is what you get for poisson distrubution) the same.

Original-CommitID: 430cbf4b78a0d12012636ff5c802a044b26ff19f