FindRange: Make it possible to search for NaNs
The labnotebook helper function FindRange is responsible for finding a
given sweep (val) in the labnotebook.

The very first test pulse has sweep NaN and currently the
GetLastSetting* family of functions do not support searching for the
testpulse data of the very first testpulse which has sweep == NaN.

Fix that and add a test for it as well.

Original-CommitID: 49bf71ab33236cc2fa0993d5cb8aeab4af6602ec