FindRange: Avoid erroring out in edge case
Since b07dd1b5 (FindRange: Fix an edge case with sweep rollback and the
entry source types, 2017-05-12) we check the source type as well.

This can lead to the situation that we have the settting
(indizesSetting exists) and the entrySourceType (indizesSourceType
exists) but the intersection of both is empty.

This can be triggered in rare cases and leads to an Igor RTE as we try
to index the indizes wave with an index of -1.

Now we check that the intersection is valid before continuing.

Original-CommitID: e3d39dfd636b989a23642774d468eb3f63dc4632