EnsureLargeEnoughWave: Avoid enlarging minimum sized waves immediately
In most cases we call EnsureLargeEnoughWave on waves which are meant to
grow pretty large and have an initial size of MINIMUM_WAVE_SIZE.

The whole point of this function is to *minimize* the number of
redimensions on the wave.

But the current design always enlarged a wave with MINIMUM_WAVE_SIZE
size on the first call. Avoid that from now on.

Original-CommitID: 3b75c7ba7e1e905e328de76783cb9cb9c5773010