ED_AddEntryToLabnotebook: Rework interface
The existing interface allows to only write one entry per headstage at a
time.

So if you would like to add one entry for multiple headstages you would
have called ED_AddEntryToLabnotebook a couple of times.

But unfortunatly this breaks our reading logic of the labnotebook.
The labnotebook querying code currently assumes that only the last
non-empty entry of a key is of interest. Earlier entries are assumed to be
"outdated" setttings. Breaking the semantics of the labnotebook is bad
enough but in addition this means that you can not use all our standard
labnotebook querying functions, GetLastSetting etc., as these only
return the very last entry. So you could only query the setting of the
last headstage you added an entry for.

This is of course bad.

Therefore we now accept a wave with entries where you can, and have to,
set all entries of a headstage at once.

Original-CommitID: 02016548f83326aa8cabad0995daad5aaec05356