Rename DC_ControlStatusWave* and rework
Now that we know that the channel checkbox controls are always
synchronized to the GUIState wave we can get rid of DC_ControlStatusWave
for nearly all callers. Only DAP_RecordDA_EphysGuiState needs to use it.

All other callers can use DC_ControlStatusWaveCache.

The functions DC_ControlStatusWave and DC_ControlStatusWaveCache also
don't belong in MIES_DataConfiguratorITC.ipf but in MIES_PanelITC.ipf
instead.

This change reduces the number of calls to GetCheckboxState from ~1100
to ~120 on a typical DAQ.

And also for example the execution time of DAP_GetITCSampInt/SI_CalculateMinSampInterval
is reduced from
0.00676463
to
0.000616625
seconds.

And as we call this function for every oscilloscope update this is a
relevant enhancement.

Original-CommitID: fc9d4263b75a71c906751703d98763b77939e939