PA_GetAveragePulseLength: Don't error out on invalid pulse range
It is currently possible for the user to specify a pulse range where
the starting pulse is larger than the ending pulse.

This leads to an error if the pulse lengths are calculated as numPulses
is zero.

Bug present since the introduction of PPA in 32ec4325 (Add pulse
averaging, 2017-02-09).

Original-CommitID: 27b3a2f661f6ce40621ec29a57cedc73ff925d8b