Analysis Functions: Implement support for new mid sweep return type
For some experiments it is beneficial to be able to abort the sweep
prematurely to save time. In these cases the user usually wants to
repurpose the not acquired time in the sweep for the ITI.

We now support that using the new return value of
ANALYSIS_FUNC_RET_REPURP_TIME for mid sweep events. The left over sweep
time in seconds is stored in GetRepurposedSweepTime and taken into
account in RA_RecalculateITI.

Original-CommitID: c85e5982b408684019dad0adf85dda0ec1bd9331