TangoInteract/TI_ConfigureMCCforIVSCC: Avoid ugly DA_EPhys panel
Our TabControl helper which is reponsible for showing and hiding the
controls in a tab does not know about nested tabs. Therefore if you show
a control in a nested sub tab it shows up in your current tab no matter
if it should be shown or not.

We workaround that issue by changing to the "Data Acquisition" tab
and restore the old tab as well.

Original-CommitID: 86fe81295464bbf53f812ef63c416f71e8fc9420