DA_EPhys: Prevent locking a ITC device which is not present
A common cause of crashes with the ITC hardware is that after the
hardware bugged out, the user unlocks and locks the device again, but as
no hardware could be found, the whole system is then locked up.

Prevent that with checking the number of available devices. Allow to
override the check in evil mode.

Original-CommitID: 9c7772961d17212b39ed211282f38dfa0953aa97