TP MD: Fix invalid extracted chunks for special baseline values
For the multi device test pulse we extract the each single test pulse
from the whole lot of test pulses located in the ITCDataWave.

For this to work correctly we have to get the written and read single
tetst pulse indizes to be the same.

For some special combinations of parameters, 4 active headstages aka
20mus sampling interval and 31% baseline, the extracted and displayed
chunks are not correct.

The reason is that for writing the test pulse into the testpulse wave,
which is later downsampled into the ITCDataWave, we use
TP_GetTestPulseLengthInPoints(panelTitle, MIN_SAMPLING_INTERVAL_TYPE)
via DC_UpdateTestPulseWave() but for reading the chunks we use
TP_GetTestPulseLengthInPoints(panelTitle, REAL_SAMPLING_INTERVAL_TYPE)
via SCOPE_UpdateOscilloscopeData.

The problem with that approach is that the real and the minimum sampling
interval number of points might be fractional.

See for example
•print TP_GetTestPulseLengthInPoints("ITC18USB_DEV_0", MIN_SAMPLING_INTERVAL_TYPE)
  13157
•print TP_GetTestPulseLengthInPoints("ITC18USB_DEV_0", REAL_SAMPLING_INTERVAL_TYPE)
  3289
•print 3289* 4
  13156

And this off-by-one error creates the problem that the testpulse seems
to shift from left to right while the ITCDataWave is acquired.

This is a fundamtenal flaw in the current code and should be present
since forever.

The fix is to always use REAL_SAMPLING_INTERVAL_TYPE which has the
consequence that we now create the test pulse wave, GetTestPulse()
already with the real sampling interval instead of creating it with the
minimum sampling interval and downsampling it.

Original-CommitID: d6d3058730bcf3e0d835d0275daa49949adfe653