introduce BrowserSettings panel with tabcontrol
4 External Subwindows are taken to a new SidePanel with TabControl to
save Display-Space.
Dependent procedures are updated to look for new controls in the sub-panel.
A button is added to the DataBrowser to show the Main Panel without
performing any action.

DataBrowser and its Control Elements are reduced in size.
Two new horizotal guides (UGH2 and UGH1) are used to define the top and bottom limit of
the graph. The third horizonal guide (UGH0) divides the graph space for
labnotebook and sweep display by 1:3.

* use enable/disable button in side panel
* disable all buttons that belong to old functionality.
* remove obselete (duplicate) buttons in parent panel
* 3 row / 3 column format in parent panel
* redirect button functionality to new panel
* some buttons were adjusted to correct height.

Original-CommitID: 3544c6fb3e0ac49497d4255cbdfcbd47a69b64d6