check if Property OVS/AR/PA is active
Previous calls only needed WindowExists() or DataFolderExistsDFR() to
check if the property is active or not.

As long as the OVS/AR/PA status depends on the panel status this is
not changing anything in functionality.

Introduce functions to check the status. This gives us better
possibility to acutally retrieve the current status.

Original-CommitID: 5e06665c6dbc1bd17f4cfb6cccb09b356b826154