DAP_ChangeHeadStageMode: Avoid ugly DA_Ephys panel
In case we change the clamp mode of a headstage via
DAP_ChangeHeadStageMode and are currently not on the data acquisition
tab the clamp mode controls of the newly choosen clamp mode will be
shown.

This is not a functional bug but plain ugly.

Avoid that by switching to the data acquisition tab if we have to.

Inspired by 83daf4be (TangoInteract/TI_ConfigureMCCforIVSCC: Avoid ugly
DA_EPhys panel, 2017-08-17).

Original-CommitID: fe2f7476d010e9e0bd772e8757ed73ac5b535ee9