TP_Delta: Fix Steady state resistance calculation for non-default baselines
The old code assumed that the same amount of time taken for the baseline
calculation can be used for the elevated calculation as well. This works
for the default (25%) baseline as then the baseline is as long as the
elevated part. But for higher baseline values this breaks as the
evaluated elevated part is getting larger than the real pulse.

Fix that by using the same timespan, 20% of the minimum of baseline and
duration, for evaluation for the elevated part as already done for the
baselines. For very long durations we just take 5ms.

Broken since 1f7cfa65 (Make the baseline length of the testpulse
configurable, 2015-07-04).

Original-CommitID: 93a29dd8ae7daceb0e384feef7ed0f4dc2680af2