ITC_StopOngoingDAQMDHelper: Reset GUI also after a crash
While developing on MIES it can sometimes happen that DAQ is stopped but
only the dataAcqRunMode variable still says that DAQ is running.

To be able to leave that state without unlocking/locking we also reset
the GUI in that state now.

Original-CommitID: 351207053db52f60b7ee02aeca0ad34573fe6803