use device to check for update
with the device beeing stored in userdata we can substitute the
procedure with a simpler comparison.

in addition the IsDataBrowser() function is used. This function could
also check the PANELTYPE userdata flag. A comment is added to this
function.

Original-CommitID: dd5d6c2f4f8a9c7b99ca68548fbbd3da204d9abf