ITC_BkrdTPFuncMD: Extract the first chunk only when its complete
Currently we extract the very first chunk (activeChunk == 0) even if it
is not complete as e.g.

 activeChunk = max(0, floor(pointsCompletedInITCDataWave / TP_GetTestPulseLengthInPoints(panelTitle)) - 1)

gives activeChunk = 0 for pointsCompletedInITCDataWave = 1.

Fix that by skipping the max invocation and only send the data to
SCOPE_UpdateOscilloscopeData if it is larger or equal to -1.

Bug present since forever.

Original-CommitID: bb42ecdb3d3347ce7e3395194bd04d69fd28a1b7