WaveBuilder: Add stimset global seed and allow epochs to use it
Some users would like to generate the exact same noise in different
epochs. Currently this is not possible as the seeds are unique per
epoch.

Introduce a new stimset seed and allow users to choose between the
stimset seed and the epoch seed.

For this we add a new entry to SegWvType, row 97, holding the stimset
seed and WP[39] holds now the user's choice if the epoch or the stimset
seed is used. WP[39] defaults and is upgraded to 1 (epoch seed) for
backwards compatibility.

Original-CommitID: 1dfa900ba52e8270fcab98e207b0abc58bbb39d0