bugfix: better support for device is NONE
create BSP panel when device is NONE but do not activate controls if no
device is assigned.

previously an ASSERT could be triggered when the panel was already open and
the user was switching from a valid divice to a NONE device.

Original-CommitID: bd2282b0d8d7356cc4fa1c8e451d2d8525834dc3