TP settings controls: Don't stop and restart TP if DAQ is ongoing
Since d221bb4e (First implementation of wave to store DA_ephys gui
settings, 2015-08-31) we stop and restart the TP if the TP settings are
changed during TP.

This makes sense if plain TP is running, but completely borks up things
if TP during ITI is running.

Fix that by not stopping and restarting TP if DAQ is running. On the
next run the new settings are then used. This works because we store all
settings in global variables and reference only these during TP and for
writing the values into the labnotebook.

Original-CommitID: 66f25e8584b163e70a07ada1e5a29cf55262441d