RA Single Device: Don't execute the after ITI functions on abort
Since forever we execute the functions ITC_STOPTestPulseSingleDevice and
RA_Counter even if we abort DAQ during ITI.

But only since c60a7168 (Count global initializes at zero instead of
NaN, 2017-07-07) the pre DAQ value of count changed from NaN to zero.
And one consequence of that is, that the `if(count < totTrials)` part of
RA_Counter is now executed if RA_Counter is called on DAQ abort.

Since 6b92d92b (DA_Ephys/Labnotebook: Introduce a "repeated
acquisition cycle ID", 2017-07-12) we require that the Repeated
acquisition cycle ID is initialized properly, and due to all this we now
have a failed assertion on DAQ abort.

Fix that by not executing the functions on abort. Returning one from a
background task is equal to stopping it.

Original-CommitID: a82e9954dce9087bed30bbe8b9c86389157e952f