Wavebuilder Panel: Introduce builtin stimset concepts
In future MIES version why might want to ship stimsets as part of the
MIES installation itself. In order to prevent that users have
accidentally created stimsets with the same name alreay, we introduce
the concept of builtin stimsets.

A builtin stimset starts with the string "MIES_" and can only be saved
with the "Allow builtin save" is checked.

Original-CommitID: 7235e4fe0f514096e133bddbde0bf2bfc5e87757