panel initialization: avoid UpdateSweepPlot()
in the AnalysisBrowser load sweep procedure the sweepBrowserDFR for the
current SweepBrowser Graph is needed before loading data to it.
Therefore the SweepBrowser needs to be created without calling
UpdatePanel().

UpdatePanel() is called concurrently after the data load has succeeded.

Possible solution to this would be to store the relation between dfr and
graph in a "database" aka "map". This can euqlly be used for the userdata.

Original-CommitID: 8bbba271eb422b1d68e95e8636808fae8e163aac