code consistency: use byRef var
As the panelTitle changes by this function to a new name, this new name
should be directed to the caller function as well to avoid confusing
code.

This also eliminates the warning comment

The panelTitleNew variable is only used as long as the new title is not
actually set. After the new panel has the new title use panelTitle again.

Original-CommitID: 67a959fa3f463b9d6f4283099e99c6b6be674408