rename variables: unify the use of panel variables
variable with same usage should be named the same. The affected
variables are:
graph, panelName, panelTitle, graphOrPanelName, panel, extPanel

The panel or graph names are unified to use the following names:
* win : unspecified Window or Subwindow
* graph: unspecified graph (in SB equals mainGraph)
* mainPanel: DB main Panel
* mainGraph: DB main Graph
* lbPanel : LabNotebook Panel
* lbGraph : LabNotebook Graph
* bsPanel : BrowserSettings Panel

extended Notes:

The variables win and graph should always be used if the required level
of detail is accurate for the function to not use a specified panel.

Specifying extPanel is not accurate in the context of SB and DB
as there is usually more than one external panel.

The variable panelTitle refers to the name of the panel and not the
title. it is altered to win, mainPanel or bsPanel, depending on
the required panel and level of detail.

Original-CommitID: 1635bcb88e13ed7c9cda1cddf25c4856a1d22676