Properly support "TP after DAQ" when aborting currently running DAQ
Since basically forever if you have "TP after DAQ" enabled and are using
Multi Device support you could not abort an ongoing DAQ with starting a
TP.

The reason is that at the forced abort of DAQ we call DAP_OneTimeCallAfterDAQ
and this starts, as requested, "TP after DAQ". But as we should start TP
now we try to start it twice.

With single device the issue was hidden by the try/catch in
TP_StartTestPulseSingleDevice.

With multi device you either got ITC errors or "error: Wave is in use by
preemptive thread. Can't be resized or killed."

We now stop quit the TP starting functions early if the TP is already
running.

Original-CommitID: 75fb61b674a1fc48c30ed61e6188c190d6aa79bd