raise DB/SB/BSP panel version
panels of SweepBrowser and DataBrowser have been changed. The
DataBrowser panel is now a graph is is the SweepBrowser. The settings
panels are unified in the BrowserSettings Panel. The SweepControl panel
is added to the bottom of the current graph. The Settings Panel is
DataBrowser only and is currently displayed under the sweepControls
panel.

Original-CommitID: 3ecd6a37bdbe502110a5020fad48a5d1d5559530