GetUniqueCtrlList: Tighten removal of special controls logic
The control check_DA_applyOnModeSwitch matches the wildcard pattern
"check_DA_*" which is used to determine if a control is unique or not.

Therefore this control is currently not accessible via the GUIState
wave.

Use the new mechanism of letting DAP_ParsePanelControl check if a
control is special or not.

In addition we don't record the state of the DA/AD associations and
pressure controls in the hardware tab as that is an hideous amount of
work to get right.

Original-CommitID: b87b9477fe8d61b2ac721f0b92c665d637ea448c