Add compilation testing to main UTF includes
This avoids that we try to run tests on igor code which does not
compile. Currently this hangs the build machine and we have to
*manually* quit Igor Pro.

We have to tweak the bamboo testing as well as we don't want to compile
the test includes for the release package as they don't exist there.

Original-CommitID: e5ffce4374d9e6e835d07b7c3703bb523a53d517