WB_CreateAndGetStimSet: Gracefully handle non-existing stimsets
Since 798b9bf4 (Analysis functions: Add central storage wave to make calling
them faster, 2017-10-24) we access the stimsets on every change of the
list of stimsets.

When you know select a stimset in a DA_EPhys panel and delete it in the
wavebuilder you get an assertion as the stimset, which you just
deleted, is tried to be recreated and that fails in the stimset childern
recursion.

Return know immediately if the stimset does not exist and also can not
be recreated.

Original-CommitID: 49e86d447da5f80b88b1ec89914c3d4d29b52e94