userdata folder property: cleaner solution
* PANEL_FOLDER:

In this folder all waves and
variables for the BrowserSettings Panel are stored. i.e. channel
selection waves. The OVS data is also stored there.

* DATA_FOLDER

This folder is currently only used by DB.
SB has a map to see where the data is stored.

* AR_SWEEPFOLDER

This folder holds the one-dimensional sweep waves.

bugfix: previously checkbox status was stored in the DATA_FOLDER. This
is not right as the DATA_FOLDER should only hold DATA and not PANEL
elements. Behaviour changed to store the checkbox state in PANEL_FOLDER

Original-CommitID: 8846c96889d5dde628fb805337621a9ae0181ea9