RA_CounterMD: Finish acquisition if required
Skipping sweeps to the end during ITI leaves the DA_Epys panel in a
state where DAQ is not finished.

The reason is that RA_CounterMD does not check if we are done.
Now we call RA_FinishAcquisition if needed.

Original-CommitID: 10fb7418daa3e88fb726b8b345468d3d926fd18f