userdata folder property: remove DATA_FOLDER
DATA_FOLDER property was previously stored in userdata.
this property is only used in DB and not in SB. This commit introduces a
more clean solution to handle this property.

DATA_FOLDER is a device-specific property.
Since the folder property is rarely used,
a more clean solution is to retrieve it directly from the device
if needed.

Original-CommitID: 7070b3114a5d62d649e939489c7747b18e06c444