hide BSP instead of closing
If the BSP is closed the status of the controls is permanently lost.
In addition, some tasks that rely on the readout of panelElements to
rebuilt the graph can not be performed.

The panel is hidden if the user tries to close it. Control elements
of hidden panels are accesible by control functions like ControlInfo.

The panel is completely rebuilt when pressing the Toggle Button.

Since the Toggle Button now only opens the panel, all toggle related
functions could be dropped.

Original-CommitID: 0573e0694645095a8395f6813ad580838eb65796