Pressure control: Fix manual pressure control
We have columns for the pressure settings from the data acquisition
panel in the guistate waves, but don't write into it and don't use the
querying functions when getting the values.

Now we write the pressure settings into the guistate wave and also read
them from it.

The bug manifested itself that for manual pressure not the current psi
value was used but the one set before. If you only changed the psi value
once this looked like that manual pressure did not work.

Broken since a31012bf (DA_Ephys: Replace GetSetVariable with
DAP_GetValueFromNumStateWave, 2017-11-02).

Original-CommitID: 623119d54424714367cd13e0ec35f6980829e9dc