feature: dDAQ slider
in SB, headstage slider was managed by a dropdown list. In DB by a
slider in the main panel. This commit unifys the use of the dDAQ slider.
One Procedure for DB and SB is introduced that calls UpdateGrpaph() when
the slider value changed.

Original-CommitID: e148c1427238c74cebe68c5d2804f315f01d0f9c