Add GetLastSetting*EachRAC
When working with repeated acquisition cycle (RAC) sweeps, one often
wants to know the value of a given setting for each sweep. This is
currently not straightforward to do, so we add four functions for querying
this information. One set for headstage dependent data and for for
independent data.

Original-CommitID: 4e7d7377b31467cfd5269a75c743acd13f2451b7