DAP_SwitchSingleMultiMode: Rework
We now allow to pass the caller if the state of the multi device
checkbox has changed or not. And only if it has changed we do the
storing and resetting checkbox state logic.

This makes the function also usable for other callers and not only the
GUI procedure function.

Original-CommitID: 35fd2d83cb9823f1248a4354d7ada30bf409e1f1