RA_CounterMD: Avoid activeSetCount being NaN for follower devices
When currently indexing is used with two yoked devices you get an error
in DC_CalculateChannelColumnNo that the column is NaN. The real reason
behind that is that we forgot to set activeSetCount for the follower
devices in RA_CounterMD.

We are now doing that correctly. But as getting indexing working
correctly with yoked devices requires a lot more work we will turn it
off in a future commit.

Original-CommitID: cd2caeb8dd5a8317905e1ffe0c04765ab0cb6517