NWB_AppendSweepLowLevel: Fix NWB export naming and metadata for unassociated channels and TTL channels
The current NWB export is, since forever, buggy in the following parts:
- The ElectrodeNumber and ElectrodeName is set for TTL datasets. This
  means the source attribute is currently

    Device=ITC18USB_Dev_0;Sweep=3;TTL=1;ElectrodeNumber=0;ElectrodeName=1;TTLBit=0

  but that does not make sense as a TTL channel is not attached to a headstage.
- Unassociated channels have the TTlBit entry set in the source attribute,
  if the same sweep had TTL data
- The naming of the unassociated channel groups was wrong as it included
  the TTLBit, e.g. data_00006_AD3_0 instead of data_00006_AD3.

The following lists exemplarly the changes in the NWB file for
NWB-bug-unassoc-AD-after-TTL.pxp found in the mies-test repository.
This compares the current master version against the changes in this
commit.

Dataset naming:

  /acquisition/timeseries/data_00004_AD3_0 -> /acquisition/timeseries/data_00004_AD3

  and so on.

Source attribute contents:

  Unassociated channels

    Device=ITC18USB_Dev_0;Sweep=4;AD=3;ElectrodeNumber=NaN;ElectrodeName=;TTLBit=0

    ->

    Device=ITC18USB_Dev_0;Sweep=4;AD=3;ElectrodeNumber=NaN;ElectrodeName=

  TTL channels

    Device=ITC18USB_Dev_0;Sweep=0;TTL=1;ElectrodeNumber=0;ElectrodeName=1;TTLBit=0

    ->

    Device=ITC18USB_Dev_0;Sweep=0;TTL=1;ElectrodeNumber=NaN;ElectrodeName=;TTLBit=0

This fix is to clear the channel relevant data in
IPWNB#WriteChannelParams before writing a new channel.

The NWB loading routines are not affected by this bug.

Original-CommitID: 9c1045fb5851f047763bfef66df6d83dbe31a09a