DAP_CheckHeadStage: Complain if analysis function is in a non-standard location
The current logic only check if the analysis function has the correct
signature. But we want to be also sure that the user defined it in the
correct place.

So let's check that. This also ensure that users can not set, via
editing the stimset parameter waves, the analysis prototype functions as
analysis function.

Original-CommitID: 32a15bd0d01175d269d26daafeaed44e341f0680