DA_Ephys: Track current FIFO position in global variable
Before we tracked the current fifo position during DAQ only in the wave
note of the oscilloscope update wave. This makes it inconvenient to
query for other future caller.

Now we store it in a global variable which is reset on DAQ start and
stop/abort to NaN and initialized to zero on each DC_ConfigureDataForITC
call.

The analysis functions referring to "lastFifoPos" from the Oscilloscope
data wave were also changed to use the global variable.

Original-CommitID: 3a9071b232860809964ca543e6392e16b52f0fd8