DAP_ButtonProc_AcquireData: First stop the TP before doing anything else
Since forever we did first check the settings via DAP_CheckSettings and
then stopped the testpulse if it is running.

Stopping a running testpulse implicitly creates labnotebook entries. And
these labnotebook entries were added after the PRE DAQ analysis
function was called. And due to that the PSQ_SquarePulse analysis
function could not find the stepsize entry in the post sweep event which
it added in the pre daq event.

Switch the order to support our implicit assumptions in the analysis
functions.

Original-CommitID: 8baca0674a4896293650cbed6332875df218c5e3