Databrowser/SweepBrowser: Make averaging work again
Since 6afe53ae (AddEntryIntoWaveNoteAsList: Add assertions, 2018-01-24)
we check that the added string to the wavenote does not contain any
semicolons as that messes up reading the entries back.

And the places where the list of source waves used for the average is
written into the wave note was missed.

Use "|" as wave separator, as that symbol is not allowed for non-liberal
wave names.

Original-CommitID: 4d53e71126b35348b61c3a9025cf90091cb3b46e