Rework stimset killing on save and quit
Since 336100a1 (Create stim sets in datafolder hierarchy only when
needed, 2015-07-09) we kill all stimsets on save and/or quit.

This was good back in the days with 32bit Igor. But nowaydays with 64bit
Igor we have much more RAM available.

So now we only kill the stimsets on save on 32bit Igor but not on 64bit
Igor. But we keep killing them on quit.

Original-CommitID: c3e2f1576634c727b265f1e7b2b6ac811c8618b8