MIES Installer update 32bit, auto uninstall, w w/o XOP mode
The package selection was replaced by a single option
to install with/without XOPs.

The installer also supports 32-bit Igor now, but recommends
to use 64-bit. If 64-bit Igor is found it defaults to 64-bit.

The installer looks for existing MIES installations on startup and
runs the uninstaller first if an installation is present. If no
installation is found but links to MIES procedures in the Igor
folder it still asks for manual cleanup.
The following limitations apply:

If the installer is running with user rights an uninstaller from
a previous installation with admin rights is not called *1).

If the installer is running silent and with admin rights the
windows privilege elevation request will pop up for the installer
and the uninstaller call. The uninstaller will continue silently.

If the installer is running with user rights it will check for a
user installation of MIES only in the default folder which is
%USERPROFILE%\Documents\MIES. The check is if uninstall.exe exists.

*1) This can be enabled if it makes any sense, the uninstaller would
ask for privilege elevation in such case.

Note on the Installer waiting for the Uninstaller to complete:
This can not be done with ExecWait as uninstall.exe extracts
another executable to a temporary folder which is then the actual
uninstaller. This allows to cleanup uninstall.exe in the installation.
The official recommendation for ExecWait locks uninstall.exe which
prevents a clean uninstallation.
The current approach here waits 300 ms after calling uninstall.exe
and waits then for the process of the extracted actual uninstaller to
finish. Also installer and uninstaller got different mutexes to prevent
multiple execution.

Basic MIES functionality with Igor Pro 8.0.0.6 Beta was tested.

All User option is disabled if not admin:
Previously when running the installer in user mode the
selection to install for all users was available at the
first page. A error message was shown much later when trying
to install though.
Now the option is disabled already if not admin.

In the source there are now constant defines at the beginning where
source paths for XOPs and path template extensions for Helpfiles
and HDF- package can be set.

Minor GUI adjustments for a clearer frontend.

Mutex protection against multiple execution has been renamed to
prevent conflict with execution of uninstaller from former version.

Original-CommitID: d977f565a35e18c8c2fb3cdc07d8de869e1de02c