Fix licensing issues with stock Igor Pro XOP/Helpfiles/Procedures
We are not allowed to publish XOPs/Helpfiles/Procedures which ship with
Igor Pro.

In preparation for a public release we replace the
XOPs/Helpfiles/Procedures with shortcuts to the on disc location.

Requires that Igor Pro is installed in the default location.

Original-CommitID: adaef595c4a9ed7ed2a0dabfd36dea46b6e1bc48