Analysis parameters: Add method to request the types as well
Real world experience showed that being able to request a specific
type for the analysis parameters helps a lot.

A breaking change is that the names now must be separated with commas
(,) as that is more in line who we store the entries in the stimset.

Original-CommitID: df4494c0a742d25c30b24d8761079ef93808b9f2