PGC_SetAndActivateControl: Do nothing if the checkbox is already in the desired state
An interactive user can never click a checkbox twice without changing
its state. But with the current version of PGC_SetAndActivateControl you
could do that and that is not intended and pretty wasteful if the
control procedure is expensive to call.

So let's to nothing if the checkbox is in checkbox mode and already in
the desired state.

Original-CommitID: 5ad72dfe5ead986bfe80e935b72a0d7b1e322eb8