Analysis Parameter panel: Avoid asserting out on missing value
Since ce36f708 (WBP_UpdateParameterWave: Add type of missing requested
parameters, 2018-05-17) we add the type of required parameter to the
listbox.

But we forgot to adjust the logic in WBP_ListBoxProc_AnalysisParams as
we know have to check if the value is not empty.

Original-CommitID: f85a6593338006c21a9425cbd26ef5ee06e1f755