Wave cache: Allow to operate on a non-duplicated wave
Until now we always duplicated the wave to store it in the cache and
duplicated it when returning it from the cache.

For users who wants to use the wave cache just for an intermediate free
wave storage this duplication does not make sense.

Therefore we now allow users to operate on the wave itself.

Original-CommitID: ea8051cee7bead58cfed8fedc603f1f057d648dd