Add a cache for the labnotebook queries
The current method of querying the labnotebook involves finding the row
range of the sweep number and then looking all those rows for an
matching entry.

This is quite time consuming. For the row range search we already have a
method to reuse old results via the first, last parameters of
GetLastSetting but this is on-demand only.

A typical query takes around 3ms on this developers machine.

Analyzing the ways we can speedup this search reveals two possibilities.
The first is that we can cache the first/last row range internally and
reuse if appropriate. The second is that we can cache the exact row of
an entry.

Both methods are now implemented and speedup the search to around 40mus.

We currently flush the cache if the labnotebook changed, we track that
via WaveModCountWrapper().

Original-CommitID: 082d20384958506dbfc129a674d4474826ef0567